moved test to sources to make them executable
